Ensembles

NQCDynamics.Ensembles — Module

Ensembles

This module provides the main function run_ensemble. This serves to run multiple trajectories for a given simulation type, sampling from an initial distribution.

NQCDynamics.Ensembles.MeanReduction — Type

Average the outputs over all trajectories.

NQCDynamics.Ensembles.OrderedSelection — Type

Select the initial conditions from the distribution in order.

NQCDynamics.Ensembles.OutputDissociation — Type

Output a 1 if the molecule has dissociated, 0 otherwise.

NQCDynamics.Ensembles.OutputFinal — Type

Output the end point of each trajectory.

NQCDynamics.Ensembles.OutputQuantisedDiatomic — Type

Output the vibrational and rotational quantum numbers of the final image.

NQCDynamics.Ensembles.OutputStateResolvedScattering1D — Type

Output a ComponentVector with fields reflection and transmission containing the probability of the outcome. Each index in the arrays refers to the adiabatic state.

NQCDynamics.Ensembles.RandomSelection — Type

Obtain initial conditions by randomly sampling the distribution.

NQCDynamics.Ensembles.SumReduction — Type

Sum the outputs from each trajectory.

NQCDynamics.Ensembles.run_ensemble — Method

run_ensemble(sim::AbstractSimulation, tspan, distribution;
    selection::Union{Nothing,AbstractVector}=nothing,
    output=(sol,i)->(sol,false),
    reduction=:append
    reduction::Symbol=:append,
    ensemble_algorithm=SciMLBase.EnsembleThreads(),
    algorithm=DynamicsMethods.select_algorithm(sim),
    saveat=[],
    trajectories=1,
    kwargs...
    )

Run multiple trajectories for timespan tspan sampling from distribution. The DifferentialEquations ensemble interface is used which allows us to specify functions to modify the output and how it is reduced across trajectories.

Keywords

selection should be an AbstractVector containing the indices to sample from the distribution. By default, nothing leads to random sampling.

output can be a function that transforms the DiffEq solution to an output, or a tuple of output quantities as for run_trajectory.

reduction defines how the data is reduced across trajectories. Options are :append, :mean or :sum.

ensemble_algorithm is the algorithm from DifferentialEquations which determines which form of parallelism is used.

algorithm is the algorithm used to integrate the equations of motion.

saveat mirrors the DiffEq keyword and labels the time points to save the output.

trajectories is the number of trajectories to perform.

This function wraps the EnsembleProblem from DifferentialEquations and passes the kwargs to the solve function.

NQCDynamics.Ensembles.select_reduction — Method

select_reduction(reduction::Symbol)

Converts reduction keyword into function that performs reduction.

:discard is used internally when using callbacks to save outputs instead.