NonadiabaticDistributions

NQCDynamics.NonadiabaticDistributions — Module

NonadiabaticDistributions

Module containing types for representing distributions of combined nuclear and electronic configurations.

NQCDynamics.NonadiabaticDistributions.Adiabatic — Type

Singleton type for labelling states as adiabatic.

NQCDynamics.NonadiabaticDistributions.BoltzmannVelocityDistribution — Method

BoltzmannVelocityDistribution(temperature, masses)

Obtain velocities sampled from the Boltzmann distribution for given temperature and masses.

This is effectively just a multivariate normal distribution but when combined with the DynamicalDistribution it ensure samples are generated with the correct size. If using for ring polymer dynamics, ensure the temperature is given as nbeads*temperature.

NQCDynamics.NonadiabaticDistributions.Diabatic — Type

Singleton type for labelling states as diabatic.

NQCDynamics.NonadiabaticDistributions.DynamicalDistribution — Type

DynamicalDistribution(velocity, position, size)

Sampleable that provides positions and velocities sampled from a variety of distributions.

Arguments

velocity and position provide the velocities and positions and can be any type for which select_item has been implemented.

size denotes to the size of the samples that should be produced.

Keywords

Example

using NQCDynamics.InitialConditions
using Distributions

d = DynamicalDistribution(5.0, Normal(), (1, 1))
rand(d)

# output

ComponentVector{Float64}(v = [5.0;;], r = [0.5314767537831963;;])

NQCDynamics.NonadiabaticDistributions.ElectronicDistribution — Type

ElectronicDistribution{S<:StateType} <: NonadiabaticDistribution

Abstract type for distributions of electronic degrees of freedom only.

NQCDynamics.NonadiabaticDistributions.ElectronicPopulation — Type

ElectronicPopulation{T,S} <: ElectronicDistribution{S}

Electronic distribution for representing a mixed state with non-zero population in multiple states.

NQCDynamics.NonadiabaticDistributions.FermiPopulation — Type

FermiPopulation{T,S} <: ElectronicDistribution{S}

Electronic distribution for Fermions following Fermi-Dirac distribution.

NQCDynamics.NonadiabaticDistributions.HarmonicRingPolymer — Method

HarmonicRingPolymer(ω, β, m, n_beads)

Ring polymer position distribution in a 1D harmonic potential

NQCDynamics.NonadiabaticDistributions.NonadiabaticDistribution — Type

NonadiabaticDistribution

Top level type for all nonadiabatic distributions.

These distributions contain a both electronic and nuclear degrees of freedom.

NQCDynamics.NonadiabaticDistributions.SingleState — Type

SingleState{S} <: ElectronicDistribution{S}

Electronic distribution for representing a system confined to a single state.

NQCDynamics.NonadiabaticDistributions.MomentumHarmonicWigner — Method

MomentumHarmonicWigner(ω, β)

Wigner distribution in a 1D harmonic potential for the momentum

NQCDynamics.NonadiabaticDistributions.PositionHarmonicWigner — Method

PositionHarmonicWigner(ω, β)

Wigner distribution in a 1D harmonic potential for the position

NQCDynamics.NonadiabaticDistributions.Q — Method

Quantum corrector for the Wigner distribution

NQCDynamics.NonadiabaticDistributions.VelocityHarmonicWigner — Method

VelocityHarmonicWigner(ω, β, m)

Wigner distribution in a 1D harmonic potential for the velocity