RingPolymers

NQCDynamics.RingPolymers.RingPolymerParameters — Method

Constructor for choosing specific elements to be quantum.

NQCDynamics.RingPolymers.RingPolymerParameters — Method

Constructor for the case where all nuclei are quantum.

NQCDynamics.RingPolymers.cayley_propagator — Method

cayley_propagator(beads::RingPolymerParameters{T}, dt::Real; half::Bool=true) where {T}

J. Chem. Phys. 151, 124103 (2019); doi: 10.1063/1.5120282

NQCDynamics.RingPolymers.constrain_classical_atoms! — Method

constrain_classical_atoms!(A::AbstractArray{T,3}, classical_atoms) where {T}

Make sure that the classical atoms all have the same values.

NQCDynamics.RingPolymers.find_rpa — Method

A = find_rpa(As) return the first RingPolymerArray among the arguments.

NQCDynamics.RingPolymers.get_centroid — Method

Evaluate centroid of ring polymer.

NQCDynamics.RingPolymers.get_normal_mode_transformation — Method

Creates normal mode transformation for n beads.

NQCDynamics.RingPolymers.get_spring_energy — Method

get_spring_energy(beads::RingPolymerParameters, masses, R)

Calculate the ring polymer spring potential.

NQCDynamics.RingPolymers.get_spring_matrix — Method

get_L(n_beads, mass, ω_n)

Return the Circulant symmetric matrix for the ring polymer springs.

NQCDynamics.RingPolymers.set_classical! — Method

Copy data from first index into replicas. All classical atoms have repeated values.