TimeCorrelationFunctions

NQCDynamics.TimeCorrelationFunctions — Module

TimeCorrelationFunctions

This module defines extra types that can be used as Ensemble outputs when computing time-correlation functions. It hopes to provide a minimal interface that reduces code repetition when implementing different correlation functions.

NQCDynamics.TimeCorrelationFunctions.PopulationCorrelationFunction — Type

PopulationCorrelationFunction{T,S<:AbstractSimulation} <: TimeCorrelationFunction

Output type for computing the population correlation function. The statetype determines the population type (diabatic or adiabatic). sim must also be provided to access the parameters to compute the population.

NQCDynamics.TimeCorrelationFunctions.TimeCorrelationFunction — Type

TimeCorrelationFunction

Abstract type for defining time correlation functions