DiabaticModels

NQCModels.DiabaticModels — Module

DiabaticModels

Models defined within this module subtype the DiabaticModel and provide potentials as Hermitian matrices and derivatives as arrays of Hermitian matrices.

NQCModels.DiabaticModels.AltDebyeSpectralDensity — Type

AltDebyeSpectralDensity{T} <: SpectralDensity

Standard Debye spectral density but uses an alternative discretization scheme that requires a cutoff parameter ωᵐ.

References

Najeh Rekik, Chang-Yu Hsieh, Holly Freedman, Gabriel Hanna, J. Chem. Phys. 138, 144106 (2013)

NQCModels.DiabaticModels.AnanthModelOne — Type

AnanthModelOne(a=0.01, b=1.6, c=0.005, d=1.0)

Ananth's simple avoided crossing model (similar to Tully's first model) from J. Chem. Phys. 127, 084114 (2007).

NQCModels.DiabaticModels.AnanthModelTwo — Type

AnanthModelTwo(a=0.04, b=0.01, c=0.005, d=1.0, e=0.7, f=1.6)

Ananth's asymmetric model from J. Chem. Phys. 127, 084114 (2007).

NQCModels.DiabaticModels.BosonBath — Type

BosonBath(density::SpectralDensity, N::Integer)

Bosonic bath with given spectral density.

Useful for sampling the bath uncoupled from the spin for spin-boson dynamics.

NQCModels.DiabaticModels.DebyeSpectralDensity — Type

DebyeSpectralDensity{T} <: SpectralDensity

Debye density as detailed in: Xin He, Jian Liu, J. Chem. Phys. 151, 024105 (2019)

NQCModels.DiabaticModels.DiabaticFrictionModel — Type

DiabaticFrictionModel <: LargeDiabaticModel

These models are defined identically to the LargeDiabaticModel but allocate extra temporary arrays when used with NQCDynamics.jl.

This allows for the calculation of electronic friction internally from the diabatic potential after diagonalisation and calculation of nonadiabatic couplings.

NQCModels.DiabaticModels.DiabaticModel — Type

DiabaticModel <: Model

DiabaticModels are used when a system has multiple electronic states that are presented in the diabatic representation. This is the case for the majority of model systems.

Implementation

DiabaticModels should implement:

potential(model, R)

derivative!(model, D, R)

nstates(model)

ndofs(model)

Example

In this example we create a simple 2 state, 1 dimensional diabatic model MyModel. As noted above, we implement the 4 relevant functions then evaluate the potential. Since this is a 1D model the argument R accepts a Real value.

using StaticArrays: SMatrix
using LinearAlgebra: Hermitian

struct MyModel <: NQCModels.DiabaticModels.DiabaticModel end

NQCModels.nstates(::MyModel) = 2
NQCModels.ndofs(::MyModel) = 1

function NQCModels.potential(::MyModel, R::Real) 
    V11 = R
    V22 = -R
    V12 = 1
    return Hermitian(SMatrix{2,2}(V11, V12, V12, V22))
end

function NQCModels.derivative!(::MyModel, D, R::Real)
    return Hermitian(SMatrix{2,2}(1, 0, 0, 1))
end

model = MyModel()
NQCModels.potential(model, 10)

# output

2×2 Hermitian{Int64, SMatrix{2, 2, Int64, 4}}:
 10    1
  1  -10

NQCModels.DiabaticModels.DoubleWell — Type

DoubleWell(mass=1, ω=1, γ=1, Δ=1)

Two state double well, also called the one-dimensional spin-boson model. See: J. Chem. Phys. 150, 244102 (2019)

NQCModels.DiabaticModels.ErpenbeckThoss — Type

struct ErpenbeckThoss{T<:AbstractFloat} <: DiabaticModel

1D two-state diabatic system capable of modelling a molecule adsorbed on a metal surface or a single-molecule junction.

In the two references, all of the parameters are identical except for the particle mass m and the vertical shift c applied to the ϵ₀ state. Both references modify the shift to ensure the quantum ground-state has an energy of 0 eV. Note that the mass m is specified in atomic mass units (amu) not atomic units. We calculate the offset automatically in the constructor from the Morse potential zero-point energy.

References

PHYSICAL REVIEW B 97, 235452 (2018)

J. Chem. Phys. 151, 191101 (2019)

NQCModels.DiabaticModels.GatesHollowayElbow — Type

GatesHollowayElbow()

Simple two state elbow potential from Gates and Holloway: Journal of Electron Spectroscopy and Related Phenomena, 64/65 (1993) 633-639

Has two diabatic states each comprised of the sum of a Morse and a repulsive potential. The coupling between them is an exponential function of z (distance from the surface).

NQCModels.DiabaticModels.LargeDiabaticModel — Type

LargeDiabaticModel <: DiabaticModel

Same as the DiabaticModels but uses normal Julia arrays instead of StaticArrays and must implement the inplace potential! rather than potential. This is useful when nstates is very large and StaticArrays are no longer efficient.

NQCModels.DiabaticModels.MiaoSubotnik — Type

Model MiaoSubotnik according to Miao & Subotnik, J. Chem Phys., 150, 041711 (2019).
V_0 =  0.5*m*ω^2 x^2
V_1 =  0.5*m*ω^2 (x-g)^2 + ΔG
Molecule-metal coupling:
V = sqrt(\Gamma*2*W/(2*\pi*M))
d is the minimum of the PES, d = g/2 +ΔG/(mω^2g) 
2W is the bandwidth
In the original model, m = 2000, ω = 0.0002, g = 20.6097, ΔG = -0.0038 and 
kBT=0.00095. In the paper, Γ, 2W and M were varied and tested.
The numbers given as default correspond to the one listed in the paper.
The model is defined such that h=(V_0-V_1)d^t d + V_1, however, the diabatic surfaces
that they show in the paper in Fig. 1 correspond to the picture implemented below

NQCModels.DiabaticModels.OhmicSpectralDensity — Type

OhmicSpectralDensity{T} <: SpectralDensity

Ohmic density as detailed in: Xin He, Jian Liu, J. Chem. Phys. 151, 024105 (2019)

NQCModels.DiabaticModels.OuyangModelOne — Type

OuyangModelOne(A=0.01, B=1.6, Γ=1e-4, N=10, ΔE=1.6e-2, D=1.0)

Model #1 from Ouyang and Subotnik. See also Ouyang's thesis.

NQCModels.DiabaticModels.SpinBoson — Type

SpinBoson(density::SpectralDensity, N::Integer, ϵ, Δ)

Spin boson model with N bosons with given spectral density.

References

Xin He, Jian Liu, J. Chem. Phys. 151, 024105 (2019)

NQCModels.DiabaticModels.ThreeStateMorse — Type

ThreeStateMorse()

Three state morse potential referred to as Model IA here: J. Chem. Phys. 150, 244102 (2019)

Models IB and IC retain the same functional form and need only a change of parameters.

NQCModels.DiabaticModels.TullyModelOne — Type

TullyModelOne(a=0.01, b=1.6, c=0.005, d=1.0)

Tully's simple avoided crossing model from J. Chem. Phys. 93, 1061 (1990).

NQCModels.DiabaticModels.TullyModelThree — Type

TullyModelThree(a=0.0006, b=0.1, c=0.9)

Tully's extended coupling with reflection model from J. Chem. Phys. 93, 1061 (1990).

NQCModels.DiabaticModels.TullyModelTwo — Type

TullyModelTwo(a=0.1, b=0.28, c=0.015, d=0.06, e=0.05)

Tully's dual avoided crossing model from J. Chem. Phys. 93, 1061 (1990).

NQCModels.DiabaticModels.discretize — Method

Discretize a given spectral density for N oscillators. Returns frequencies and couplings.