MACE interface

Possible output data

NQCModels.AdiabaticModels.MACE.MACEModel — Type

MACE interface with support for ensemble of models and batch size selection for potentially faster inference.

DOFs currently hardcoded at 3.

Fields

model_paths::Vector{String}: Model paths to load. If multiple paths are given, an ensemble of models is used and mean energies/forces used to propagate dynamics.

models::Vector: Loaded models.

device::Vector{String}: Device to use for inference. If a single device is given, all models will be loaded on that device. If multiple devices are given, each model will be loaded on the corresponding device. Default is "cpu".

default_dtype::Type{T}: Default data type for inference. Default is Float32.

batch_size::Union{Int,Nothing}: Batch size for inference. Default is 1.

cutoff_radius::T: Cutoff radius for the models.

last_eval_cache::MACEPredictionCache: Cache for the last evaluation.

z_table: Table of atomic numbers.

atoms::Atoms: Atoms object for the system.

cell::PeriodicCell: Periodic cell for the system.

ndofs::Int: Number of degrees of freedom in the system.

mobile_atoms::Vector{Int}: Indices of mobile atoms in the system.

NQCModels.AdiabaticModels.MACE.MACEModel — Method

MACEModel(
atoms::Atoms,
cell::AbstractCell,
model_paths::Vector{String};
device::Union{String,Vector{String}}="cpu",
default_dtype::Type=Float32,
batch_size::Int=1,
mobile_atoms::Vector{Int}=collect(1:length(atoms)),

) modelpaths::Vector{String}, device::Union{String, Vector{String}}="cpu", defaultdtype::Type=Float64, batch_size::Int=1, atoms, cell, )

Interface to MACE machine learning interatomic potentials with support for ensemble of models and batch size selection for potentially faster inference.

Arguments

atoms: Atoms object for the system. predict! can be called with a Vector of different Atoms objects to evaluate different structures.

cell: Cell object for the system. predict! can be called with a Vector of different Cell objects to evaluate different structures.

model_paths::Vector{String}: Model paths to load. If multiple paths are given, an ensemble of models is used and mean energies/forces used to propagate dynamics.

device::Union{String, Vector{String}}: Device to use for inference. If a single device is given, all models will be loaded on that device. If multiple devices are given, each model will be loaded on the corresponding device. Default is "cpu".

default_dtype::Type: Default data type for PyTorch (usually Float32)

batch_size::Int: Batch size for inference. Can be set to nothing to always adapt batch size to the number of structures provided. Ring-polymer methods benefit from this.

mobile_atoms::Vector{Int}: Indices of mobile atoms in the system. Default is all atoms.

NQCModels.AdiabaticModels.MACE.MACEPredictionCache — Type

Cache which stores the results of the last evaluation of the MACE model. This is used to speed up concurrent energy and force evaluations, as well as to give access to ensemble predictions.

Use e.g. get_energy_mean(MACEModel.last_eval_cache) to access results of the last prediction made by the model.

NQCModels.AdiabaticModels.MACE.get_energy_ensemble — Method

get_energy_ensemble(mace_cache::MACEPredictionCache)

Returns the potential energy evaluated by each model in the ensemble in units of Hartree.

NQCModels.AdiabaticModels.MACE.get_energy_mean — Method

get_energy_mean(mace_cache::MACEPredictionCache)

Returns the mean potential energy of the structures stored in the evaluation cache.

Energy is returned in units of Hartree.

NQCModels.AdiabaticModels.MACE.get_energy_variance — Method

get_energy_variance(mace_cache::MACEPredictionCache)

Returns the variance of the potential energy in Hartree².

NQCModels.AdiabaticModels.MACE.get_forces_ensemble — Method

get_forces_ensemble(mace_cache::MACEPredictionCache)

Returns the forces evaluated by each model in the ensemble in units of Hartree/Bohr.

NQCModels.AdiabaticModels.MACE.get_forces_mean — Method

get_forces_mean(mace_cache::MACEPredictionCache)

Returns the mean forces of the structures stored in the evaluation cache. Forces are returned in units of Hartree/Bohr.

NQCModels.AdiabaticModels.MACE.get_forces_variance — Method

get_forces_variance(mace_cache::MACEPredictionCache)

Returns the force variance of the structures stored in the evaluation cache. Forces are returned in units of Hartree²/Bohr².

NQCModels.AdiabaticModels.MACE.mace_configuration_from_nqcd_configuration — Method

mace_configuration_from_nqcd_configuration(
atoms::Atoms,
cell::Union{InfiniteCell, PeriodicCell},
R::AbstractMatrix,

)

Converter into a single mace.data.utils.Configuration to make use of MACE's data loading.

NQCModels.AdiabaticModels.MACE.predict! — Method

predict!(
mace_interface::MACEModel,
atoms::Union{Vector{<:Atoms}, Atoms},
R::Vector{<:AbstractMatrix},
cell::Union{Vector{<:AbstractCell}, AbstractCell},
)

Evaluate the MACE model on a set of structures if they haven't already been evaluated. The results are stored in mace_interface.last_eval_cache.

Use methodswith(MACEPredictionCache) to see what data is provided from the results

Arguments

mace_interface: MACEModel to evaluate with.

atoms: Atoms object for the structures passed for evaluation. Can be either a vector of different Atoms objects or a single Atoms object.

R: Vector of structures to evaluate.

cell: Cell object for the structures passed for evaluation. Can be either a vector of different Cell objects or a single Cell object.

NQCModels.AdiabaticModels.MACE.predict — Method

predict(
mace_interface::MACEModel,
atoms::Union{Vector{<:Atoms}, Atoms},
R::Vector{<:AbstractMatrix},
cell::Union{Vector{<:AbstractCell}, AbstractCell},
)

Evaluate the MACE model on a set of structures if they haven't already been evaluated. The results are returned as a MACEPredictionCache.

Use methodswith(MACEPredictionCache) to see what data can be provided from the results.

Arguments

mace_interface: MACEModel to evaluate with.

atoms: Atoms object for the structures passed for evaluation. Can be either a vector of different Atoms objects or a single Atoms object.

R: Vector of structures to evaluate.

cell: Cell object for the structures passed for evaluation. Can be either a vector of different Cell objects or a single Cell object.

NQCModels.derivative! — Method

NQCModels.derivative(model::MACEModel, atoms::Atoms, R::AbstractArray{T,3}, cell::Union{InfiniteCell, PeriodicCell})

This variant of NQCModels.derivative can make use of batch evaluation to speed up inference for multiple structures.

NQCModels.derivative — Method

NQCModels.derivative(model::MACEModel, atoms::Atoms, R::Vector{<:AbstractMatrix}, cell::Union{InfiniteCell, PeriodicCell})

This variant of NQCModels.derivative can make use of batch evaluation to speed up inference for multiple structures.

NQCModels.potential — Method

NQCModels.potential(model::MACEModel, atoms::Atoms, R::Vector{<:AbstractMatrix}, cell::AbstractCell)

This variant of NQCModels.potential can make use of batch evaluation to speed up inference for multiple structures.