DynamicsMethods

NQCDynamics.DynamicsMethods — Module

This module contains functions and types necessary for performing nonadiabatic molecular dynamics.

Dynamics is performed using DifferentialEquations.jl. As such, this module is centered around the implementation of the functions necessary to integrate the dynamics.

For deterministic Hamiltonian methods, the central function is DynamicsMethods.motion!, which is the inplace form of the function to be integrated by DifferentialEquations.jl.

Further, methods that have discontinuities, such as surface hopping, use the callback interface provided by DifferentialEquations.jl.

NQCDynamics.DynamicsMethods.Method — Type

Each type of dynamics subtypes Method which is passed to the AbstractSimulation as a parameter to determine the type of dynamics desired.

NQCDynamics.DynamicsMethods.DynamicsVariables — Method

DynamicsVariables(::AbstractSimulation, args...)

For each dynamics method this function is implemented to provide the variables for the dynamics in the appropriate format.

By default, DynamicsVariables is set up for the classical case and takes sim, v, r as arguments and returns a ComponentVector(v=v, r=r) which is used as a container for the velocities and positions during classical dynamics.

NQCDynamics.DynamicsMethods.create_problem — Method

Provides the DEProblem for each type of simulation.

NQCDynamics.DynamicsMethods.get_callbacks — Method

Select the default callbacks for this simulation type.

NQCDynamics.DynamicsMethods.motion! — Function

motion!(du, u, sim, t)

As per DifferentialEquations.jl, this function is implemented for each method and defines the time derivatives of the DynamicalVariables.

We require that each implementation ensures du and u are subtypes of DynamicalVariables and sim subtypes AbstractSimulation.

NQCDynamics.DynamicsMethods.run_trajectory — Method

run_trajectory(u0, tspan::Tuple, sim::AbstractSimulation;
               output=(:u,), saveat=[], algorithm=select_algorithm(sim), kwargs...)

Solve a single trajectory starting from u0 with a timespan tspan for the simulation sim.

Keyword arguments

output specifies the quantities that should be saved during the dynamics simulation. The options for this keyword are any of the functions found in src/DynamicsMethods/output.jl.

The rest of the keywords are the usual arguments for the solve function from DifferentialEquations.jl. It is possible to use Unitful quantities for any of the arguments since these are automatically converted to atomic units internally.

Output

The function returns a Table from TypedTables.jl with columns for time t and every quantity specified in the output tuple.

NQCDynamics.DynamicsMethods.select_algorithm — Method

Choose a default algorithm for solving the differential equation.

ClassicalMethods

NQCDynamics.DynamicsMethods.ClassicalMethods.Classical — Type

Classical <: DynamicsMethods.Method

Type for performing classical molecular dynamics.

sim = Simulation{Classical}(Atoms(:H), Harmonic())

# output

Simulation{Classical}:
  Atoms{Float64}([:H], [1], [1837.4715941070515])
  Harmonic{Float64, Float64, Float64}
  m: Float64 1.0
  ω: Float64 1.0
  r₀: Float64 0.0
  dofs: Int64 1

NQCDynamics.DynamicsMethods.ClassicalMethods.Langevin — Type

Type for performing Langevin molecular dynamics.

using Unitful
sim = Simulation{Langevin}(Atoms(:H), Free(); γ=2.5, temperature=100u"K")

# output

Simulation{Langevin{Float64}}:
  Atoms{Float64}([:H], [1], [1837.4715941070515])
  Free(1)

NQCDynamics.DynamicsMethods.ClassicalMethods.MDEF — Type

\[dr = v dt\\ dv = -\Delta U/M dt - \Gamma v dt + \sigma \sqrt{2\Gamma} dW\]

$\Gamma$ is the friction tensor with units of inverse time. For thermal dynamics we set $\sigma = \sqrt{kT / M}$, where $T$ is the electronic temperature.

This is integrated using the BAOAB algorithm where the friction "O" step is performed in the tensor's eigenbasis. See src/dynamics/mdef_baoab.jl for details.

NQCDynamics.DynamicsMethods.ClassicalMethods.ThermalLangevin — Type

Type for performing Langevin ring polymer molecular dynamics.

Currently there are separate types for classical and ring polymer versions of Langevin dynamics but they should be combined. The reason they are not at the moment is that they use different integration algorithms and require slightly different fields.

using Unitful
RingPolymerSimulation{ThermalLangevin}(Atoms(:H), Free(), 10; γ=0.1, temperature=25u"K")

# output

RingPolymerSimulation{ThermalLangevin{Float64}}:
 
  Atoms{Float64}([:H], [1], [1837.4715941070515])
 
  Free(1)
  with 10 beads.

NQCDynamics.DynamicsMethods.ClassicalMethods.acceleration! — Method

f1 in DifferentialEquations.jl docs.

NQCDynamics.DynamicsMethods.ClassicalMethods.acceleration! — Method

acceleration!(dv, v, r, sim::Simulation{MDEF,<:DiabaticFrictionCalculator}, t)

Sets acceleration due to ground state force when using a DiabaticFrictionModel.

NQCDynamics.DynamicsMethods.ClassicalMethods.friction! — Method

friction!(g, r, sim, t)

Evaluates friction tensor

MappingVariableMethods

NQCDynamics.DynamicsMethods.MappingVariableMethods.NRPMD — Type

NRPMD{T} <: DynamicsMethods.Method

Nonadiabatic ring polymer molecular dynamics Uses Meyer-Miller-Stock-Thoss mapping variables for electronic degrees of freedom and ring polymer formalism for nuclear degrees of freedom.

RingPolymerSimulation{NRPMD}(Atoms(:H), DoubleWell(), 10)

# output

RingPolymerSimulation{NRPMD{Float64}}:
 
  Atoms{Float64}([:H], [1], [1837.4715941070515])
 
  DoubleWell{Int64, Int64, Int64, Int64}
  mass: Int64 1
  ω: Int64 1
  γ: Int64 1
  Δ: Int64 1
 
  with 10 beads.

NQCDynamics.DynamicsMethods.MappingVariableMethods.eCMM — Type

eCMM{T} <: DynamicsMethods.Method

References

Xin He, Zhihao Gong, Baihua Wu, Jian Liu (2021)

Xin He, Baihua Wu, Zhihao Gong, Jian Liu (2021)

SurfaceHoppingMethods

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods — Module

SurfaceHoppingMethods

Implementation for surface hopping methods.

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.FSSH — Type

FSSH{T} <: SurfaceHopping

Type for fewest-switches surface hopping

Simulation{FSSH}(Atoms(:H), Free())

# output

Simulation{FSSH{Float64}}:
  Atoms{Float64}([:H], [1], [1837.4715941070515])
  Free(1)

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.SurfaceHopping — Type

Abstract type for all surface hopping methods.

Surface hopping methods follow the structure set out in this file. The nuclear and electronic variables are propagated by the motion! function. The surface hopping procedure is handled by the HoppingCallback which uses the functions check_hop! and execute_hop! as its condition and affect!.

To add a new surface hopping scheme, you must create a new struct and define methods for evaluate_hopping_probability!, select_new_state, and rescale_velocity!.

See fssh.jl for an example implementation.

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.calculate_a — Method

calculate_a(sim::AbstractSimulation{<:SurfaceHopping}, coupling::AbstractMatrix)

Equation 40 from Sharon Hammes‐Schiffer, John C. Tully (1994).

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.calculate_b — Method

calculate_b(coupling::AbstractMatrix, velocity::AbstractMatrix)

Equation 41 from Sharon Hammes‐Schiffer, John C. Tully (1994).

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.evaluate_hopping_probability! — Method

evaluate_hopping_probability!(sim::Simulation{<:FSSH}, u, dt)

Evaluates the probability of hopping from the current state to all other states

Implementation

σ is Hermitan so the choice σ[m,s] or σ[s,m] is irrelevant; we take the real part.

'd' is skew-symmetric so here the indices are important.

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.evaluate_hopping_probability! — Method

This function should set the field sim.method.hopping_probability.

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.extract_nonadiabatic_coupling — Method

extract_nonadiabatic_coupling(coupling, new_state, old_state)

Extract the nonadiabatic coupling vector between states new_state and old_state

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.perform_rescaling! — Method

perform_rescaling!(
    sim::AbstractSimulation{<:SurfaceHopping}, velocity, velocity_rescale, d
)

Equation 33 from Sharon Hammes‐Schiffer, John C. Tully (1994).

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.rescale_velocity! — Method

rescale_velocity!(sim::AbstractSimulation{<:FSSH}, u)::Bool

Rescale the velocity in the direction of the nonadiabatic coupling.

References

Sharon Hammes‐Schiffer, John C. Tully (1994)

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.rescale_velocity! — Method

This function should modify the velocity and return a Bool that determines whether the state change should take place.

This only needs to be implemented if the velocity should be modified during a hop.

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.select_new_state — Method

This function should return the desired state determined by the probability. Should return the original state if no hop is desired.

NQCDynamics.DynamicsMethods.SurfaceHoppingMethods.unpack_states — Method

unpack_states(sim)

Get the two states that we are hopping between.

EhrenfestMethods

NQCDynamics.DynamicsMethods.EhrenfestMethods.AbstractEhrenfest — Type

Abstract type for Ehrenfest method.

NQCDynamics.DynamicsMethods.EhrenfestMethods.Ehrenfest — Type

Ehrenfest{T} <: AbstractEhrenfest

Ehrenfest molecular dynamics. Classical molecular dynamics where the force is derived by averaging contributions from multiple electronic states.

Simulation{Ehrenfest}(Atoms(:H), DoubleWell())

# output

Simulation{Ehrenfest{Float64}}:
  Atoms{Float64}([:H], [1], [1837.4715941070515])
  DoubleWell{Int64, Int64, Int64, Int64}
  mass: Int64 1
  ω: Int64 1
  γ: Int64 1
  Δ: Int64 1

IntegrationAlgorithms

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.MInt — Type

MInt <: OrdinaryDiffEq.OrdinaryDiffEqAlgorithm

Second order symplectic momentum integral algorithm.

Reference

J. Chem. Phys. 148, 102326 (2018)

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.RingPolymerMInt — Type

RingPolymerMInt <: OrdinaryDiffEq.OrdinaryDiffEqAlgorithm

Second order symplectic momentum integral algorithm applied to NRPMD.

Reference

J. Chem. Phys. 148, 102326 (2018)

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.get_C_propagator — Method

Get the C propagator for the mapping variables.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.get_D_propagator — Method

Get the D propagator for the mapping variables.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.get_gamma — Method

Get the Γ variable used to calculate the nuclear propagators.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.get_mapping_nuclear_force — Method

Get the force due to the mapping variables.

Equivalent to this but doesn't allocate: return 0.5 * (q'Eq + p'Ep) - q'Fp

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.get_mapping_nuclear_force — Method

Get the force due to the mapping variables.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.get_xi — Method

Get the Ξ variable used to calculate the nuclear propagators.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.set_C_propagator! — Method

Get the C propagator for the mapping variables.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.set_D_propagator! — Method

Get the D propagator for the mapping variables.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.set_gamma! — Method

Get the Γ variable used to calculate the nuclear propagators.

NQCDynamics.DynamicsMethods.IntegrationAlgorithms.set_xi! — Method

Get the Ξ variable used to calculate the nuclear propagators.