DynamicsOutputs

Here are all the functions that you can specify in the output tuple when using run_dynamics. To add more, simply add a new function in the DynamicsOutputs module.

NQCDynamics.DynamicsOutputs — Module

DynamicsOutputs

Defines a set of functions that can be used to calculate outputs for dynamics simulations.

NQCDynamics.DynamicsOutputs.OutputDissociation — Type

Output a 1 if the molecule has dissociated, 0 otherwise.

NQCDynamics.DynamicsOutputs.OutputQuantisedDiatomic — Type

Output the vibrational and rotational quantum numbers of the final image.

NQCDynamics.DynamicsOutputs.OutputStateResolvedScattering1D — Type

Output a ComponentVector with fields reflection and transmission containing the probability of the outcome. Each index in the arrays refers to the adiabatic state.

NQCDynamics.DynamicsOutputs.OutputAdiabaticPopulation — Method

OutputAdiabaticPopulation(sol, i)

Output the adiabatic population at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputCentroidPosition — Method

OutputCentroidPosition(sol, i)

Output the position of the ring polymer centroid at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputCentroidVelocity — Method

OutputCentroidVelocity(sol, i)

Output the velocity of the ring polymer centroid at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputDiabaticPopulation — Method

OutputDiabaticPopulation(sol, i)

Output the diabatic population at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputDiscreteState — Method

OutputDiscreteState(sol, i)

Output the discrete state variable at each timestep during the trajectory. This is used for surface hopping simulations and returns the variable that determines the currently occupied adiabatic state.

Requires that the dynamics variable has a field state.

Use OutputDiabaticPopulation or OutputAdiabaticPopulation to get the population estimators.

NQCDynamics.DynamicsOutputs.OutputDynamicsVariables — Method

OutputDynamicsVariables(sol, i)

Output all of the dynamics variables at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputFinal — Method

Output the end point of each trajectory.

NQCDynamics.DynamicsOutputs.OutputKineticEnergy — Method

OutputKineticEnergy(sol, i)

Evaluate the classical kinetic energy at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputMappingMomentum — Method

OutputMappingMomentum(sol, i)

Output the momentum mapping variable at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputMappingPosition — Method

OutputMappingPosition(sol, i)

Output the position mapping variables at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputPosition — Method

OutputPosition(sol, i)

Output the position at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputPotentialEnergy — Method

OutputPotentialEnergy(sol, i)

Output the adiabatic potential energy at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputQuantumSubsystem — Method

OutputQuantumSubsystem(sol, i)

Output the quantum subsystem at each timestep during the trajectory. Usually this will refer to a wavefunction or density matrix but will depend on the particular dynamics method.

NQCDynamics.DynamicsOutputs.OutputSurfaceHops — Method

Output the total number of surface hops during the trajectory

NQCDynamics.DynamicsOutputs.OutputTotalAdiabaticPopulation — Method

OutputTotalAdiabaticPopulation(sol, i)

Output the total adiabatic population at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputTotalDiabaticPopulation — Method

OutputTotalDiabaticPopulation(sol, i)

Output the total diabatic population at eah timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputTotalEnergy — Method

OutputTotalEnergy(sol, i)

Evaluate the classical Hamiltonian at each timestep during the trajectory.

NQCDynamics.DynamicsOutputs.OutputVelocity — Method

OutputVelocity(sol, i)

Output the velocity at each timestep during the trajectory.

Internals