DynamicsOutputs

Here are all the functions that you can specify in the output tuple when using run_trajectory. To add more, simply add a new function in the DynamicsOutputs module.

NQCDynamics.DynamicsOutputs — Module

DynamicsOutputs

Infrastructure for saving quantities during trajectories.

Defines all available options that can be used within the output tuple along with the functions that perform the saving operation.

NQCDynamics.DynamicsOutputs.adiabatic_population — Method

Evaluate the adiabatic population

NQCDynamics.DynamicsOutputs.centroid_position — Method

Get the position of the ring polymer centroid

NQCDynamics.DynamicsOutputs.force — Method

Get the force

NQCDynamics.DynamicsOutputs.friction — Method

Evaluate the friction. This is used for MDEF only.

NQCDynamics.DynamicsOutputs.hamiltonian — Method

Evaluate the classical Hamiltonian

NQCDynamics.DynamicsOutputs.kinetic — Method

Evaluate the classical kinetic energy

NQCDynamics.DynamicsOutputs.population — Method

Evaluate the diabatic population

NQCDynamics.DynamicsOutputs.position — Method

Get the position

NQCDynamics.DynamicsOutputs.potential — Method

Evaluate the potential from the model

NQCDynamics.DynamicsOutputs.quantum_subsystem — Method

Get the quantum subsystem of the dynamics variables. Requires that DynamicsUtils.get_quantum_subsystem is implemented for your chosen method.

NQCDynamics.DynamicsOutputs.state — Method

Get the currently occupied state from the dynamics variables. Requires that the dynamics variable has a field state. Currently this is for surface hopping methods only. Use population or adiabatic_population for most other methods.

NQCDynamics.DynamicsOutputs.u — Method

Get all the dynamics variables. This is the default

NQCDynamics.DynamicsOutputs.velocity — Method

Get the velocity

Internals

NQCDynamics.DynamicsOutputs.OutputSaver — Method

(output::OutputSaver)(u, t, integrator)

Evaluates every function listed in output.function_names and returns all the results.

NQCDynamics.DynamicsOutputs.OutputSaver — Method

OutputSaver(function_names::NTuple{N, Symbol}) where {N}

Used to obtain the outputs for all functions given in function_names.

NQCDynamics.DynamicsOutputs.centroid_velocity — Method

Get the velocity of the ring polymer centroid

NQCDynamics.DynamicsOutputs.create_saving_callback — Method

create_saving_callback(quantities::NTuple{N, Symbol}; saveat=[]) where {N}

Get the SavingCallback that will populate saved_values with the result obtained by evaluating the functions provided in function_names.

NQCDynamics.DynamicsOutputs.evaluate_output! — Method

Used to recursively evaluate every function for the OutputSaver.

See here for a description of why it is written like this: https://stackoverflow.com/questions/55840333/type-stability-for-lists-of-closures