InitialConditions

NQCDynamics.InitialConditions — Module

InitialConditions

Functions and types for generating initial conditions for simulations.

ThermalMonteCarlo

NQCDynamics.InitialConditions.ThermalMonteCarlo.run_advancedhmc_sampling — Method

run_advancedhmc_sampling(sim::Simulation, r, n_samples;
    target_acceptance=0.5, kwargs...)

Perform Hamiltonian Monte Carlo sampling for the simulation sim using AdvancedHMC.jl.

NQCDynamics.InitialConditions.ThermalMonteCarlo.run_advancedmh_sampling — Method

run_advancedhmc_sampling(sim, r, steps, σ; move_ratio=0.0, internal_ratio=0.0)

Sample the configuration space for the simulation sim starting from r.

Total number of steps is given by steps and σ is the dictionary of step sizes for each species.

move_ratio defaults to 0.0 and denotes the fraction of system moved each step. If move_ratio = 0, every degree of freedom is moved at each step. If move_ratio = 1, then nothing will happen. Experiment with this parameter to achieve optimal sampling.

internal_ratio works as for move_ratio but for the internal modes of the ring polymer.

QuantisedDiatomic

NQCDynamics.InitialConditions.QuantisedDiatomic — Module

QuantisedDiatomic

This module exports two user facing functions:

generate_configurations Creates a set of velocities and positions for diatomic molecule with specified vibrational ν and rotational J quantum numbers.

quantise_diatomic Obtains vibrational ν and rotational J quantum numbers for a diatomic molecule with a given set of velocities and positions.

The central concept of this module is the EBK procedure which is nicely detailed here: Andrew J. Larkoski, David G. Ellis, Lorenzo J. Curtis (2006)

Inspired by VENUS96: William L Hase, Ronald J Duchovic, Xiche Hu, Andrew Komornicki, Kieran F Lim, Da-hong Lu, Gilles H Peslherbe, Kandadai N Swamy, SR Vande Linde, Antonio Varandas, others (1996)

NQCDynamics.InitialConditions.QuantisedDiatomic.apply_random_rotation! — Method

Randomly rotate each column of two 3*N matrix, same rotation for all columns.

NQCDynamics.InitialConditions.QuantisedDiatomic.calculate_diatomic_energy — Method

calculate_diatomic_energy(model::AdiabaticModel, bond_length::Real;
    height=10, normal_vector=[0, 0, 1])

Returns potential energy of diatomic with bond_length at height from surface.

Orients molecule parallel to the surface at the specified height, the surface is assumed to intersect the origin. This requires that the model implicitly provides the surface, or works fine without one.

NQCDynamics.InitialConditions.QuantisedDiatomic.calculate_force_constant — Method

calculate_force_constant(sim::Simulation, surface::SurfaceParameters)

Evaluate energy for different bond lengths and identify force constant.

This uses LsqFit.jl to fit a morse potential with optimised minimum.

NQCDynamics.InitialConditions.QuantisedDiatomic.combine_slab_and_molecule — Method

combine_slab_and_molecule(atom_indices, molecule, slab)

Revert the transformation separate_slab_and_molecule

NQCDynamics.InitialConditions.QuantisedDiatomic.configure_diatomic — Method

Randomly orient molecule in space for a given bond length and radial momentum

NQCDynamics.InitialConditions.QuantisedDiatomic.find_total_energy — Method

find_total_energy(sim, ν, J, surface)

Returns the energy associated with the specified quantum numbers

NQCDynamics.InitialConditions.QuantisedDiatomic.generate_configurations — Method

generate_configurations(sim, ν, J; samples=1000, height=10, normal_vector=[0, 0, 1],
    translational_energy=0, direction=[0, 0, -1], position=[0, 0, height])

Generate positions and momenta for given quantum numbers

translational_energy, direction and position specify the kinetic energy in a specific direction with the molecule placed with centre of mass at position.

Keyword arguments height and normal_vector become relevant if the potential requires specific placement of the molecule. These allow the molecule to be placed at a distance height in the direction normal_vector when performing potential evaluations.

NQCDynamics.InitialConditions.QuantisedDiatomic.quantise_diatomic — Method

quantise_diatomic(sim::Simulation, v::Matrix, r::Matrix;
    height=10, normal_vector=[0, 0, 1])

Quantise the vibrational and rotational degrees of freedom for the specified positions and velocities

When evaluating the potential, the molecule is moved to height in direction normal_vector. If the potential is independent of centre of mass position, this has no effect. Otherwise, be sure to modify these parameters to give the intended behaviour.

NQCDynamics.InitialConditions.QuantisedDiatomic.select_random_bond_lengths — Method

Pick a random bond length and corresponding radial momentum that matches the radial probability distribution.

Uses rejection sampling: https://en.wikipedia.org/wiki/Rejection_sampling

NQCDynamics.InitialConditions.QuantisedDiatomic.separate_slab_and_molecule — Method

separate_slab_and_molecule(atom_indices, r)

Get the coordinates of the molecule and slab separately.

MetropolisHastings

NQCDynamics.InitialConditions.MetropolisHastings — Module

MetropolisHastings

Sampling of the initial conditions using the Metropolis-Hastings Markov chain Monte Carlo method.

Included within is the ability to sample the canonical distribution for adiabatic classical and ring polymer systems.

Usage involves creating an instance of an AbstractSystem{MonteCarlo} and calling run_monte_carlo_sampling.

NQCDynamics.InitialConditions.MetropolisHastings.MonteCarlo — Type

Parameters for Monte carlo simulations.

NQCDynamics.InitialConditions.MetropolisHastings.MonteCarloOutput — Type

Container for storing simulation quantities

NQCDynamics.InitialConditions.MetropolisHastings.acceptance_probability — Method

acceptance_probability(system::AbstractSystem{MonteCarlo})

Return the Metropolis-Hastings acceptance probability.

NQCDynamics.InitialConditions.MetropolisHastings.apply_random_perturbation! — Method

apply_random_perturbation!(system::AbstractSystem{MonteCarlo}, R::AbstractMatrix, atom::Integer)

Randomly perturb the xyz coordinates of a single atom.

NQCDynamics.InitialConditions.MetropolisHastings.assess_proposal! — Method

assess_proposal!(system::AbstractSystem{MonteCarlo}, Rᵢ, Rₚ, output, i)

Update the energy, check for acceptance, and update the output.

NQCDynamics.InitialConditions.MetropolisHastings.propose_centroid_move! — Method

propose_centroid_move!(system::RingPolymerSystem{MonteCarlo}, Rᵢ::Array{T, 3}, Rₚ::Array{T, 3}) where {T<:AbstractFloat}

Propose a move for the ring polymer centroid for one atom.

NQCDynamics.InitialConditions.MetropolisHastings.propose_move! — Method

propose_move!(system::System{MonteCarlo}, Rᵢ::Matrix{T}, Rₚ::Matrix{T}) where {T<:AbstractFloat}

Propose simple cartesian move for a single atom.

NQCDynamics.InitialConditions.MetropolisHastings.propose_normal_mode_move! — Method

propose_normal_mode_move!(system::RingPolymerSystem{MonteCarlo}, Rᵢ::Array{T, 3}, Rₚ::Array{T, 3}) where {T}

Propose a move for a single normal mode for a single atom.

NQCDynamics.InitialConditions.MetropolisHastings.run_main_loop! — Method

run_main_loop!(system::System{MonteCarlo}, Rᵢ::Matrix, Rₚ::Matrix, output::MonteCarloOutput)

Main loop for classical systems.

NQCDynamics.InitialConditions.MetropolisHastings.run_main_loop! — Method

Main loop for ring polymer systems.

NQCDynamics.InitialConditions.MetropolisHastings.run_monte_carlo_sampling — Method

run_monte_carlo_sampling(sim::AbstractSimulation, monte::MonteCarloParameters, R0)

Perform Monte Carlo sampling for the system defined by the sim and monte parameters.

From the initial positions specified R0 the system will be explored using the Metropolis-Hastings algorithm.

NQCDynamics.InitialConditions.MetropolisHastings.write_output! — Method

write_output!(output::MonteCarloOutput, Rᵢ::AbstractArray, energy::AbstractFloat, i::Integer)

Store the current configuration and associated energy.